Allow any invariant register in Memset. Memset is now optimized for float array/typedarray Added tests to memop invariant implementation

+++ lib/Runtime/Library/TypedArray.h
@@ -341,7 +341,7 @@ namespace Js
             {
                 const size_t byteSize = sizeof(TypeName) * newLength;
                 Assert(byteSize >= newLength); // check for overflow
-                memset(typedBuffer + newStart, typedValue, byteSize);
+                memset(typedBuffer + newStart, (int)typedValue, byteSize);
             }
             else
             {
